[#17] refactor : 컴포넌트 분리
