+++ src/pages/KeywordPage.js
@@ -1,6 +1,7 @@
-import React from 'react';
+import React,{useState} from 'react';
 import SideNavbar from 'components/SideNavbar';
-import KeywordList from 'components/Keyword/KeywordList';
+import KeywordFilterBar from 'components/Keyword/KeywordFilter';
+import KeywordHeader from 'components/Keyword/KeywordHeader';
 import styled from 'styled-components';
 
 const Container = styled.div`
@@ -9,10 +10,14 @@ const Container = styled.div`
 
 
 const KeywordPage = () => {
+
+  const [isToggle,setIsToggle] = useState(false);
+
   return (
     <Container>
       <SideNavbar/>
-      <KeywordList/>
+      <KeywordHeader toggle={isToggle} title={'키워드 알림'}/>
+      <KeywordFilterBar toggle={isToggle}/>
     </Container>
   );
 };
